+++ lib/acme/resources/authorization.rb
@@ -1,9 +1,9 @@
 class Acme::Resources::Authorization
   HTTP01 = Acme::Resources::Challenges::HTTP01
   DNS01 = Acme::Resources::Challenges::DNS01
-  DVSNI01 = Acme::Resources::Challenges::DVSNI01
+  TLSSNI01 = Acme::Resources::Challenges::TLSSNI01
 
-  attr_reader :domain, :status, :http01, :dns01, :dvsni01
+  attr_reader :domain, :status, :http01, :dns01, :tlssni01
 
   def initialize(client, response)
     @client = client
@@ -18,7 +18,7 @@ class Acme::Resources::Authorization
       case attributes.fetch('type')
       when 'http-01' then @http01 = HTTP01.new(@client, attributes)
       when 'dns-01' then @dns01 = DNS01.new(@client, attributes)
-      when 'tls-sni-01' then @dvsni01 = DVSNI01.new(@client, attributes)
+      when 'tls-sni-01' then @tlssni01 = TLSSNI01.new(@client, attributes)
       else
         # no supported
       end
