+++ lib/acme/resources/challenges.rb
@@ -1,5 +1,5 @@
 module Acme::Resources::Challenges; end
 require 'acme/resources/challenges/base'
 require 'acme/resources/challenges/http01'
-require 'acme/resources/challenges/dvsni01'
+require 'acme/resources/challenges/tlssni01'
 require 'acme/resources/challenges/dns01'
